digital and analogue operations in CIM architectures
The combination of digital and analogue operations in in-store computing architectures allows for more efficient data processing. For example, analogue computing can be used to initially process sensor data, and then digital computing can be used to further analyse and make decisions on that data. This hybrid approach takes advantage of the strengths of both types of computing to improve overall system performance and energy efficiency.